split filters for cleaner code, adjustred inline-ness
