+++ runtime-transaction/src/lib.rs
@@ -1,6 +1,7 @@
 #![cfg_attr(feature = "frozen-abi", feature(min_specialization))]
 #![allow(clippy::arithmetic_side_effects)]
 
+mod builtin_programs_filter;
 pub mod compute_budget_instruction_details;
 mod compute_budget_program_id_filter;
 pub mod instructions_processor;
